Reorganized file structure and seperated CSS styling into seperate modules for each component.
