+++ src/App.module.css
@@ -1,0 +1,31 @@
+@import url('https://fonts.googleapis.com/css2?family=Bad+Script&family=Caveat&family=Fascinate&family=Marck+Script&family=Nunito:wght@300&family=Pacifico&family=Patrick+Hand&display=swap');
+
+
+*{  
+    margin: 0px;
+    padding: 0px;
+    font-family: 'Nunito', sans-serif;
+}
+
+body {
+    background-color: rgb(0,159,183);
+}
+
+h1, h2, h3, h4 {
+    font-family: 'Pacifico', cursive;
+    color: rgb(38, 49, 52);
+    margin: 5px;
+}
+
+link {
+    text-decoration: none;
+}
+
+ul {
+    list-style-type: none;
+}
+
+input {
+    background-color: rgb(230,230,234);
+    border-radius: 4px;
+}
