+++ src/Pages/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ToDoContainer from '../components/ToDoContainer/ToDoContainer';
-import Header from '../components/Header.js'
+import Header from '../components/Header/Header.js'
 function HomePage () {
   // tableID and tableName states for API calls and page elements 
   const tableID = "tbln7S0L5bCcELBkR";
