+++ src/components/AddToDoForm/AddToDoForm.module.css
@@ -1,0 +1,9 @@
+.AddToDoButton {
+    background-color: rgb(255,112,112);
+    color: rgb(54,73,78);
+    font-size: medium;
+    width: 4em;
+    margin: 4px 5px 0px 5px;
+    border: 'none';
+    border-radius: 12px;
+ }
