+++ src/components/Header/Header.js
@@ -1,13 +1,14 @@
 import React from 'react'
 import {Link}  from 'react-router-dom';
-import style from './ToDoList.module.css';
+import style from './Header.module.css';
 
 const Header= () => {
 
     return (
     <div className= {style.Header}>
-        <Link to="/" className= {style.NavItem}>Home</Link>
+        <Link to="/Home" className= {style.NavItem}>Home</Link>
         <Link to="/goals" className= {style.NavItem}>Goals</Link>
+        <Link to="/" className= {style.NavLogout}>Logout</Link>
     </div>)
     
 }
