+++ src/components/InputWithLabel/InputWithLabel.module.css
@@ -1,0 +1,5 @@
+.InputBar {
+    margin: 2px 6px 0px 6px;
+    padding: 2px;
+    width: 150px;
+}
