--- src/components/ToDoList.module.css
@@ -1,104 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Bad+Script&family=Caveat&family=Fascinate&family=Marck+Script&family=Nunito:wght@300&family=Pacifico&family=Patrick+Hand&display=swap');
-
-
-*{  
-    margin: 0px;
-    padding: 0px;
-    font-family: 'Nunito', sans-serif;
-}
-
-.Root {
-    color: rgb(38, 49, 52);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-width: 400px;
-}
-
-.Header {
-    color: rgb(38, 49, 52);
-    background-color: rgb(254,215,102);
-    padding: 8px;
-    margin: 0px;
-    height: 3em;
-    justify-content: right;
-    position: sticky;
-}
-
-.Container {
-    margin: 5px;
-    padding: 10px;
-    overflow: hidden;
-}
-
-.ListItem {
-    margin: 5px;
-    padding: 5px;
-}
-
-.NavItem {
-    margin: 5px 10px 0px 10px;
-    font-family: 'Pacifico', cursive;
-    color: rgb(38, 49, 52);
-    text-decoration: none;
-    font-size:x-large;
-}
-
-.NavItem  {
-
-}
-
-.AddToDoButton {
-   background-color: rgb(255,112,112);
-   color: rgb(54,73,78);
-   font-size: medium;
-   width: 4em;
-   margin: 0px 5px 0px 5px;
-   border: 'none';
-   border-radius: 12px;
-}
-
-.InputBar {
-    margin: 2px 6px 0px 6px;
-    padding: 2px;
-}
-
-.RemoveButton {
-    border: 'none';
-    width: 18px;
-    height: 18px;
-    margin: 0px 0px 0px 20px;
-}
-
-.SortToggle {
-    border: 'none';
-    margin: 8px 25% 0px 25%;
-    background-color: rgb(255,112,112);
-    color: rgb(54,73,78);
-    border-radius: 8px;
-    font-size: medium;
-    width: 6em;
-}
-
-body {
-    background-color: rgb(0,159,183);
-}
-
-h1, h2, h3, h4 {
-    font-family: 'Pacifico', cursive;
-    color: rgb(38, 49, 52);
-    margin: 5px;
-}
-
-link {
-    text-decoration: none;
-}
-
-ul {
-    list-style-type: none;
-}
-
-input {
-    background-color: rgb(230,230,234);
-    border-radius: 4px;
-}
