+++ src/components/ToDoListItem/ToDoListItem.module.css
@@ -1,0 +1,13 @@
+.ListItem {
+    margin: 4px;
+    padding: 2px;
+}
+
+.RemoveButton {
+    border: 'none';
+    width: 18px;
+    height: 18px;
+    margin: 0px 0px 0px 10%;
+    
+    justify-content: space-between;
+}
